Move assertions to verify step
