+++ zavod/zavod/exporters/statistics.py
@@ -39,58 +39,8 @@ def get_country_facets(countries: Dict[str, int]) -> List[Any]:
     return facets
 
 
-def compare_threshold(value: int, comparison: Comparison, threshold: int) -> bool:
-    match comparison:
-        case Comparison.GTE:
-            return value >= threshold
-        case Comparison.LTE:
-            return value <= threshold
-        case _:
-            raise ValueError(f"Unknown comparison: {comparison}")
-
-
-def get_value(stats: Dict[str, Any], assertion: Assertion) -> Optional[int]:
-    match assertion.metric:
-        case Metric.ENTITY_COUNT:
-            match assertion.filter_attribute:
-                case "schema":
-                    items = stats["things"]["schemata"]
-                    filter_key = "name"
-                case "country":
-                    items = stats["things"]["countries"]
-                    filter_key = "code"
-                case _:
-                    raise ValueError(
-                        f"Unknown filter attribute: {assertion.filter_attribute}"
-                    )
-            items = [i for i in items if i[filter_key] == assertion.filter_value]
-            if len(items) != 1:
-                return None
-            return cast(int, items[0]["count"])
-        case Metric.COUNTRY_COUNT:
-            return len(stats["things"]["countries"])
-        case _:
-            raise ValueError(f"Unknown metric: {assertion.metric}")
-
-
-def check_assertion(
-    context: Context, stats: Dict[str, Any], assertion: Assertion
-) -> None:
-    value = get_value(stats, assertion)
-    if value is None:
-        context.log.warning(f"Value not found for assertion {assertion}")
-        return
-    if not compare_threshold(value, assertion.comparison, assertion.threshold):
-        context.log.warning(f"Assertion failed for value {value}: {assertion}")
-
-
-class StatisticsExporter(Exporter):
-    TITLE = "Dataset statistics"
-    FILE_NAME = STATISTICS_FILE
-    MIME_TYPE = "application/json"
-
-    def setup(self) -> None:
-        super().setup()
+class Statistics(object):
+    def __init__(self):
         self.entity_count = 0
         self.last_change: Optional[str] = None
         self.schemata: Set[str] = set()
@@ -104,7 +54,7 @@ class StatisticsExporter(Exporter):
         self.target_countries: Dict[str, int] = defaultdict(int)
         self.target_schemata: Dict[str, int] = defaultdict(int)
 
-    def feed(self, entity: Entity) -> None:
+    def observe(self, entity: Entity) -> None:
         self.entity_count += 1
         self.schemata.add(entity.schema.name)
         for prop in entity.iterprops():
@@ -128,8 +78,8 @@ class StatisticsExporter(Exporter):
             else:
                 self.last_change = max(self.last_change, entity.last_change)
 
-    def finish(self) -> None:
-        stats = {
+    def as_dict(self):
+        return {
             "last_change": self.last_change,
             "schemata": list(self.schemata),
             "properties": list(self.qnames),
@@ -147,9 +97,20 @@ class StatisticsExporter(Exporter):
             },
         }
 
-        for assertion in self.dataset.assertions:
-            check_assertion(self.context, stats, assertion)
 
+class StatisticsExporter(Exporter):
+    TITLE = "Dataset statistics"
+    FILE_NAME = STATISTICS_FILE
+    MIME_TYPE = "application/json"
+
+    def setup(self) -> None:
+        super().setup()
+        self.stats = Statistics()
+
+    def feed(self, entity: Entity) -> None:
+        self.stats.observe(entity)
+
+    def finish(self) -> None:
         with open(self.path, "wb") as fh:
-            write_json(stats, fh)
+            write_json(self.stats.as_dict(), fh)
         super().finish()
